Setting up the project structure
