+++ filtered_news.csv
@@ -1,6 +1,6 @@
 Идентификатор,Название,Дата публикации,Рейтинг
-51404,Вышел дебютный клип One Fish Orchestra «Пинабалерина»,1726502258,0.88
-51419,Правила игры: рецензия на фильм «Виды доброты» Йоргоса Лантимоса,1726590847,0.5
-51416,"Стало известно, когда выйдет сериал о Гарри Поттере",1726584634,0.5
-51413,Кинолог назвал самые популярные осенние имена для собак,1726577183,0.5
-51410,Был объявлен шорт-лист Букеровской премии,1726573815,0.5
+51733,Путешествие без границ с VK Видео: блоги о необычных маршрутах и лайфхаках для путешествий,1729175470,0.73
+51523,Племя изгоев: рецензия на фильм «Граф Монте-Кристо»,1727444480,0.62
+51744,Кинематографисты готовят ремейк «Невероятных приключений итальянцев в России»,1729181899,0.5
+51743,Появился новый трейлер финального сезона «Аркейна»,1729180585,0.5
+51742,В тизере нового «Уоллеса и Громита» у главных героев неприятности,1729180418,0.5
