+++ news_report.csv
@@ -1,0 +1,64 @@
+Идентификатор,Название,Дата публикации,Рейтинг
+51744,Кинематографисты готовят ремейк «Невероятных приключений итальянцев в России»,1729181899,0.5
+51743,Появился новый трейлер финального сезона «Аркейна»,1729180585,0.5
+51742,В тизере нового «Уоллеса и Громита» у главных героев неприятности,1729180418,0.5
+51741,Разработаны новые правила поступления в вузы для будущих учителей,1729177662,0.5
+51733,Путешествие без границ с VK Видео: блоги о необычных маршрутах и лайфхаках для путешествий,1729175470,0.73
+51740,Вышел дебютный трейлер байопика про Александра Пушкина,1729174744,0.5
+51739,Открыт приём заявок на премию для фуд-блогеров,1729174362,0.5
+51737,Умер бывший солист One Direction Лиам Пейн,1729164126,0.5
+51732,Собака забралась на самую высокую пирамиду Египта,1729095150,0.5
+51731,У ЕГЭ не будет альтернативы,1729091924,0.5
+51730,Героев «Гарри Поттера» превратили в котов,1729091779,0.5
+51729,Водительские права предложили выдавать с 16 лет,1729088668,0.5
+51609,"Go по работам: основательница благотворительного проекта «Путь домой» про осознанную помощь, ценность жизни и ответственность за тех, кого мы приручили",1729087396,0.5
+51725,В России вырос спрос на тарологов и астрологов,1729073941,0.5
+51722,Райану Рейнольдсу и Хью Джекману предложили стать ведущими «Оскара»,1729008606,0.5
+51721,Над Россией летит самая яркая комета за 13 лет,1729005951,0.5
+51716,Спектакль окончен: рецензия на фильм «Джокер: Безумие на двоих»,1728992057,0.5
+51713,Роман с призраком: рецензия на фильм «Сидони в Японии»,1728988882,0.5
+51709,В России могут запретить выступления животных в цирке,1728922007,0.5
+51707,Идёт работа над новым сериалом про Штирлица,1728920976,0.5
+51706,Котят нашли в коробке с чипсами,1728916403,0.5
+51704,В октябре можно будет наблюдать суперлуние,1728914532,0.5
+51702,Одноклассники запускают уникальное научпоп-шоу про садоводство,1728907764,0.5
+51698,Измени своё лицо: рецензия на фильм «Другой человек»,1728818890,0.5
+51690,У «Теории большого взрыва» появится спин-офф,1728657965,0.5
+51693,В Риме за посещение фонтана Треви будут брать плату,1728657216,0.5
+51691,Tesla выпустит беспилотные такси,1728657032,0.5
+51688,Вышел ещё один трейлер «Конклава» с Рэйфом Файнсом,1728649999,0.5
+51679,Британцы снимут спин-офф «Гордости и предубеждения»,1728645579,0.5
+51678,В России выпустили книгу о пет-френдли культуре,1728645460,0.5
+51683,Северное сияние наблюдали по всей России,1728643216,0.5
+51676,Составили рейтинг лучших хорроров всех времён,1728567207,0.5
+51673,ВКонтакте отмечает 18-летие,1728564918,0.5
+51665,Планета против: рецензия на фильм «Говорит Земля!»,1728548446,0.5
+51662,В Японии выпустили самый маленький кубик Рубика в истории,1728489302,0.5
+51661,Вышел дебютный трейлер байопика про Боба Дилана,1728485891,0.5
+51660,Роман «Йеллоуфейс» Ребекки Куанг экранизируют,1728484649,0.5
+51659,В новом трейлере «Моаны 2» герои противостоят богине,1728483965,0.5
+51657,Объявлена дата премьеры экранизации «Ста лет одиночества»,1728483584,0.5
+51647,Турецкие пастухи требуют от фермеров оплачивать интернет для просмотра TikTok,1728399458,0.5
+51643,Москва возглавила рейтинг самых матерящихся городов страны,1728397841,0.5
+51634,Нейросеть помогла сделать трейлер игровой версии «Принцессы Мононоке»,1728315009,0.5
+51633,Шоу «Натальная карта» запустило сервис с предсказаниями,1728314796,0.5
+51632,Энн Хэтэуэй подтвердила участие в третьей части «Дневников принцессы»,1728314643,0.5
+51631,В России может появиться студенческая ипотека,1728312093,0.5
+51630,В октябре ожидаются два звездопада,1728311203,0.5
+51629,В России планируют ввести штрафы за оскорбление учителей,1728310966,0.5
+51620,Над Россией зажглось северное сияние,1728296475,0.5
+51615,Смотрим новый трейлер фильма «Паддингтон в Перу»,1728055367,0.5
+51614,Большой театр теперь продаёт билеты только по паспортам,1728052593,0.5
+51613,В Нью-Йорке крысам будут давать контрацептивы для сокращения их численности,1728050685,0.5
+51612,Эминем станет дедушкой в следующем году,1728048000,0.5
+51289,В Питере — строить бизнес. История агентства MUSICTOTHE,1727950665,0.5
+51563,Кровавое боди: рецензия на фильм «Субстанция»,1727789838,0.5
+51523,Племя изгоев: рецензия на фильм «Граф Монте-Кристо»,1727444480,0.62
+51290,"Go по работам: юрист по интеллектуальной собственности про народные представления о профессии, молодую музыкальную индустрию и Гарри Поттера",1727437016,0.5
+51503,Утопия по-русски: рецензия на сериал «Исправление и наказание»,1727184467,0.5
+51468,Социальный хоррор: рецензия на фильм «Холли»,1727173125,0.5
+51394,"Go по работам: кастинг-директор Татьяна Хречкова о режиссёрском видении, новых лицах в кино и будущем кастинга",1726838450,0.5
+51237,"Go по работам: интерьерный фотограф о секретах отличного фото, правильной съёмке и силовых тренировках",1726834664,0.5
+51435,Радости смерти: рецензия на фильм «Битлджус Битлджус»,1726677557,0.5
+51419,Правила игры: рецензия на фильм «Виды доброты» Йоргоса Лантимоса,1726590847,0.5
+51401,Старая (не)добрая комедия: рецензия на фильм «Ненасытные люди»,1726496880,0.5
